Aborted atttempt to make beam runner hosted in docker
This can't work yet because the service binds explicitly to localhost, and there is no way to port-forward that out of the docker container. If the main service allowed the binding interface to be overridden by flags, this would work. There would then be some fun with setting up volumes for disk access too most likely or all of the job output would stay trapped inside the container.