Updated terraform config to support modern tooling
 - Updated to latest version of the google module
 - Drop the beta provider
 - Extracted the variables out to a separate file as per modern terraform standards
 - Update to latest version of Trillian
 - Cleaner delete operations
 - Update docs about temporary instances
 - Use recent version of GKE
 - Use safer version of granting IAM permissions

This at least allows the terraform apply command to complete. Unforunately the kubernetes configs are massively out of date and given that the etcd-operator project is archived (https://github.com/coreos/etcd-operator), it could be that much larger changes are required to bring this up to modern standards. OTOH, in a case of very weird timing, this project has just appeared: https://github.com/aenix-io/etcd-operator.